% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-tabela.r
\name{new_tabela}
\alias{new_tabela}
\title{Construtor Da Classe \code{tabela}}
\usage{
new_tabela(
  nome,
  campos,
  uri,
  tipo_arquivo,
  particoes = NULL,
  descricao = NULL,
  no_master = FALSE
)
}
\arguments{
\item{nome}{string indicando o nome da tabela}

\item{campos}{lista de objetos \code{campo}. Veja \code{\link{new_campo}} para mais detalhes}

\item{uri}{caminho do diretorio ou bucket contendo o(s) arquivo(s) que compoe(m) a tabela}

\item{tipo_arquivo}{extensao do(s) arquivo(s) que compoe(m) a tabela}

\item{particoes}{opcional, vetor de nomes dos campos pelos quais a tabela e particionada}

\item{descricao}{opcional, breve descricao da tabela e o que contem}

\item{no_master}{booleano indicando se tabelas mestras devem ser construidas ou nao, mesmo que
haja particionamento de tabela. Veja Detalhes}
}
\value{
objeto descritivo de uma tabela
}
\description{
Funcao para geracao de objetos representativos das tabelas em um determinado banco de dados
}
\details{
Esta funcao faz parte do backend do pacote \code{dbinterface}. Atraves dela e possivel
especificar a estrutura de qualquer tabela regular, definindo seus campos e correspondentes
tipos, particoes e outros atributos.

\code{no_master} existe apenas para uso com o morgana. Neste caso, e necessario criar uma conexao
mock para reutilizar parte da infraestrutura de composicao de queries, mas não existe necessidade
de tabelas mestras. Ainda mais em bancos muito extensos em que sua construcao pode ser demorada e
causar time-out da conexao.
}
\examples{

# contruindo uma tabela nao particionada a partir do dado exemplo do pacote
# utilizando um pequeno subset de suas colunas para simplificar o exemplo
tab <- dbinterface:::new_tabela(
    nome = "subbacias",
    campos = list(
        dbinterface:::new_campo("posto", "int"),
        dbinterface:::new_campo("nome", "string"),
        dbinterface:::new_campo("codigo", "string")
    ),
    uri = system.file("extdata/cpart_parquet/subbacias/", package = "dbinterface"),
    tipo_arquivo = ".parquet.gzip"
)

# construindo uma tabela particionada
tab <- dbinterface:::new_tabela(
    nome = "vazoes",
    campos = list(
        dbinterface:::new_campo("data", "date"),
        dbinterface:::new_campo("codigo", "string"),
        dbinterface:::new_campo("vazao", "float")
    ),
    uri = system.file("extdata/cpart_parquet/vazoes/", package = "dbinterface"),
    particoes = c("codigo"),
    tipo_arquivo = ".parquet.gzip",
    descricao = "Dados de entrada de vazao observada"
)

}
\seealso{
Construtor externo \code{\link{schema2tabela}}. Funcao geral para acessar dados das
tabelas \code{\link{getfromdb}}
}
