% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes-escrita.r
\name{switch_writer_func}
\alias{switch_writer_func}
\title{Seleciona Funcao Escritora Por Tipo E Fonte De Arquivo}
\usage{
switch_writer_func(extensao, s3 = FALSE)
}
\arguments{
\item{extensao}{string indicando a extensao de arquivo, preferencialmente precedida de "."}

\item{s3}{booleano indidicano se o arquivo esta no s3 ou nao}
}
\value{
funcao cujo primeiro argumento e o objeto a ser salvo e o segundo o caminho
(local ou s3) do arquivo
}
\description{
Para uma dada \code{extensao} de arquivo e sua origem (local ou s3) retorna uma funcao para
escrita do mesmo
}
\examples{

# para escrita    de um json
fun1 <- dbinterface:::switch_writer_func("json", FALSE)
identical(body(fun1)[[1]], jsonlite::write_json)

# para escrita de um csv
fun2 <- dbinterface:::switch_writer_func(".csv", FALSE)
identical(body(fun2)[[1]], data.table::fwrite)

}
