% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbinterface.r
\docType{package}
\name{dbinterface}
\alias{dbinterface-package}
\alias{dbinterface}
\title{Pacote \code{dbinterface}}
\description{
Backend Para Acesso A Bancos De Dados \code{mock} E Relacionais
}
\details{
Este pacote fornece funcionalidade backend para interface com bancos \code{mock}
localizados em um diretorio local ou em um bucket S3, alem de bancos relacionais tradicionais.
A versao atual suporta particionamento de dados, porem apenas operacoes de leitura
estao implementadas. Criacao de novas tabelas e escrita nas existentes
serao adicionadas em versoes futuras.
}
\author{
\strong{Maintainer}: Lucas Khenayfis \email{lucaskhenayfis@gmail.com}

}
